start to  move stuff around
