+++ pysages/backends/openmm.py
@@ -18,25 +18,14 @@ from pysages.backends.snapshot import Box, Snapshot
 from jax.dlpack import from_dlpack as asarray
 from jaxlib.xla_extension import DeviceArray as JaxArray
 
+from .core import ContextWrapper
 
-class ContextWrapper:
+class ContextWrapperOpenMM(ContextWrapper):
     def __init__(self, context, force):
+        super().__init__(context)
         self.view = force.view(context)
-        self.context = context
         self.synchronize = self.view.synchronize
 
-
-class Sampler:
-    def __init__(self, method_bundle, bias):
-        snapshot, initialize, update = method_bundle
-        self.snapshot = snapshot
-        self.state = initialize()
-        self.update_from = update
-        self.bias = bias
-    #
-    def update(self):
-        self.state = self.update_from(self.snapshot, self.state)
-        self.bias(self.snapshot, self.state)
 
 
 def is_on_gpu(view: ContextView):
